+++ vllm/distributed/eplb/rebalance_execute.py
@@ -19,9 +19,6 @@ from torch.distributed import (
     get_global_rank,
 )
 
-from vllm.distributed.parallel_state import (
-    get_ep_group,
-)
 from vllm.logger import init_logger
 
 logger = init_logger(__name__)
@@ -510,131 +507,6 @@ async def transfer_layer(
         ep_group=ep_group,
     )
     return is_unchanged, is_received_locally, recv_metadata
-
-
-def find_cycles(old_expert_indices: torch.Tensor, new_expert_indices: torch.Tensor):
-    """
-    Find all cycles in the permutation from old to new.
-    Args:
-        old_expert_indices: Shape (num_physical_experts).
-        new_expert_indices: Shape (num_physical_experts).
-    """
-    n = len(old_expert_indices)
-    visited = [False] * n
-    cycles = []
-
-    for start_pos in range(n):
-        if (
-            visited[start_pos]
-            or old_expert_indices[start_pos] == new_expert_indices[start_pos]
-        ):
-            continue
-
-        cycle = []
-        pos = start_pos
-
-        while not visited[pos]:
-            visited[pos] = True
-            cycle.append(pos)
-
-            # Which expert is now at this position?
-            expert = new_expert_indices[pos].item()
-            # Where was this expert before?
-            next_pos = (old_expert_indices == expert).nonzero()[0].item()
-
-            pos = next_pos
-
-        if len(cycle) > 1:
-            cycles.append(cycle)
-
-    return cycles
-
-
-def apply_transfer_cap(
-    old_expert_indices: torch.Tensor,
-    new_expert_indices: torch.Tensor,
-    max_transfers: int,
-) -> None:
-    """
-    Cap transfers by undoing complete cycles. Modifies new in-place.
-    Args:
-        old_expert_indices: Shape (num_physical_experts).
-        new_expert_indices: Shape (num_physical_experts).
-    """
-    cycles = find_cycles(old_expert_indices, new_expert_indices)
-
-    # Calculate total transfers
-    total_transfers = sum(len(cycle) for cycle in cycles)
-
-    if total_transfers <= max_transfers:
-        return
-
-    # Sort cycles by length (undo smallest first)
-    cycles.sort(key=len)
-
-    transfers_remaining = total_transfers
-
-    for cycle in cycles:
-        # Undo this cycle
-        for pos in cycle:
-            new_expert_indices[pos] = old_expert_indices[pos]
-
-        transfers_remaining -= len(cycle)
-
-        # If we're now under the cap, stop
-        if transfers_remaining <= max_transfers:
-            break
-
-
-def cap_num_transfers(
-    old_global_expert_indices: torch.Tensor,
-    new_global_expert_indices: torch.Tensor,
-    num_tensors_per_expert: int,
-    max_num_transfers: int = 1000,
-) -> None:
-    """
-    Detects if the total number of transfer operations exceeds
-    max_num_transfers. If it does, then apply_transfer_cap will undo transfers
-    until the number of transfers is below max_num_transfers.
-
-    Args:
-        old_global_expert_indices: Shape (num_moe_layers, num_physical_experts).
-        new_global_expert_indices: Shape (num_moe_layers, num_physical_experts).
-        num_weights_per_expert: Number of tensors in each layer
-        associated with a single expert.
-        max_num_transfers: Maximum number of total transfers allowed.
-    """
-    num_moe_layers = old_global_expert_indices.shape[0]
-    ep_group = get_ep_group().device_group
-    ep_rank = ep_group.rank()
-
-    is_main_rank = ep_rank == 0
-    if is_main_rank:
-        num_transfers = []
-        for layer in range(num_moe_layers):
-            old_layer_indices = old_global_expert_indices[layer]
-            new_layer_indices = new_global_expert_indices[layer]
-            num_expert_transfers = (old_layer_indices != new_layer_indices).sum().item()
-            num_transfers.append(num_expert_transfers * num_tensors_per_expert)
-        logger.info("NUM TRANSFERS BEFORE: %s", num_transfers)
-    for layer in range(num_moe_layers):
-        old_layer_indices = old_global_expert_indices[layer]
-        new_layer_indices = new_global_expert_indices[layer]
-        num_expert_transfers = (old_layer_indices != new_layer_indices).sum().item()
-        total_num_transfers = num_expert_transfers * num_tensors_per_expert
-        if total_num_transfers >= max_num_transfers:
-            max_expert_transfers = max_num_transfers // num_tensors_per_expert
-            apply_transfer_cap(
-                old_layer_indices, new_layer_indices, max_expert_transfers
-            )
-    if is_main_rank:
-        num_transfers = []
-        for layer in range(num_moe_layers):
-            old_layer_indices = old_global_expert_indices[layer]
-            new_layer_indices = new_global_expert_indices[layer]
-            num_expert_transfers = (old_layer_indices != new_layer_indices).sum().item()
-            num_transfers.append(num_expert_transfers * num_tensors_per_expert)
-        logger.info("NUM TRANSFERS AFTER: %s", num_transfers)
 
 
 def rearrange_expert_weights_inplace(
